run one diamond for the one universal contig

+++ ppanini/tools/ppanini_gene_caller.py
@@ -136,7 +136,7 @@ def main():
     index_name = utilities.index(new_contig_file)
     
     # reads alignment using Bowtie2
-    alignment_file = utilities.alignment(args.fastq, index_name)
+    gene_alignment_file = utilities.alignment(args.fastq, index_name)
     
     # make directory for featureCounts abundance output
     config.temp_dir = temp_dir+'/featureCounts_output/'
@@ -144,44 +144,68 @@ def main():
     
     # gene abundance using featureCounts
     config.temp_dir = temp_dir
-    abundance_file = utilities.abundance(genes_file_gff, alignment_file)
-    
-    # Run diamond
-    alignment_file = utilities.diamond_alignment(genes_file_faa, args.uniref )
-    
-    # Infer abundance for sufficient hits to  uniref90 and no_hits
-    hits, no_hits, uniref_gene_map = utilities.Infer_aligmnets(alignment_file, config.temp_dir)
-    
-    # write map file to the output
-    #with open(config.output_folder+'/'+config.file_basename+'_hits.txt' , 'w') as foo:
-    #    foo.writelines(hits) #header
-    
-    # select sequence for insufficient hits
-    no_hits_genes_faa = utilities.select_sequnces(genes_file_faa, no_hits, output_name = '_no_hits.faa')
-    
-    # select sequence for sufficient hits 
-    hits_genes_faa = utilities.select_sequnces(genes_file_faa, hits, output_name = '_hits.faa')
-    
-    
-    # move the three main output under main output folder from temp files
-    # if there is more than one contig and the prodigal outputs ahvn't been produces (first sample run)
+    abundance_file = utilities.abundance(genes_file_gff, gene_alignment_file)
     shutil.move(abundance_file, config.output_folder+'/'+os.path.basename(os.path.normpath(abundance_file)))
-    
-    shutil.move(no_hits_genes_faa, config.output_folder+'/no_hits/'+ config.file_basename+ '_no_hits.faa')
-    shutil.move(hits_genes_faa, config.output_folder+'/hits/'+ config.file_basename+ '_hits.faa')
-    shutil.move(no_hits, config.output_folder+'/no_hits/'+ config.file_basename+ '_no_hits.txt')
-    shutil.move(hits, config.output_folder+'/hits/'+ config.file_basename+ '_hits.txt')
-    shutil.move(uniref_gene_map, config.output_folder+'/hits/'+ config.file_basename+ '_uniref_gene_map.txt')
-    if not os.path.isfile(config.output_folder+"/prodigal.gff"):
-        shutil.move(genes_file_gff, config.output_folder+'/prodigal.gff')
-    if not os.path.isfile(config.output_folder+"/prodigal.faa"):
-        shutil.move(genes_file_faa, config.output_folder+'/prodigal.faa')
-    print ("Main output files for ppanini_press are written in: \n%s\n%s\n%s\n%s\n%s")% (config.output_folder+'/'+os.path.basename(os.path.normpath(abundance_file)),
-                                        config.output_folder+'/no_hits/'+ config.file_basename+ '_no_hits.faa',
-                                        config.output_folder+'/no_hits/'+ config.file_basename+ '_no_hits.txt',
-                                        config.output_folder+'/hits/'+ config.file_basename+ '_hits.faa',
-                                        config.output_folder+'/hits/'+ config.file_basename+ '_hits.txt',
-                                        config.output_folder+'/hits/'+ config.file_basename+ '_uniref_gene_map.txt')
+    # if each sample has its own contig
+    if not args.one_contig:
+        # Run diamond
+        uniref_alignment_file = utilities.diamond_alignment(genes_file_faa, args.uniref )
+        
+        # Infer abundance for sufficient hits to  uniref90 and no_hits
+        hits, no_hits, uniref_gene_map = utilities.Infer_aligmnets(uniref_alignment_file, config.temp_dir)
+        
+        # select sequence for insufficient hits
+        no_hits_genes_faa = utilities.select_sequnces(genes_file_faa, no_hits, output_name = '_no_hits.faa')
+        
+        # select sequence for sufficient hits 
+        hits_genes_faa = utilities.select_sequnces(genes_file_faa, hits, output_name = '_hits.faa')
+    else:
+        if not os.path.isfile(config.output_folder+'/no_hits/no_hits.txt') or not os.path.isfile(config.output_folder+'/hits/uniref_gene_map'): 
+            # Run diamond
+            uniref_alignment_file = utilities.diamond_alignment(genes_file_faa, args.uniref )
+            
+            # Infer abundance for sufficient hits to  uniref90 and no_hits
+            hits, no_hits, uniref_gene_map = utilities.Infer_aligmnets(uniref_alignment_file, config.temp_dir)
+            
+            # select sequence for insufficient hits
+            no_hits_genes_faa = utilities.select_sequnces(genes_file_faa, no_hits, output_name = '_no_hits.faa')
+            
+            # select sequence for sufficient hits 
+            hits_genes_faa = utilities.select_sequnces(genes_file_faa, hits, output_name = '_hits.faa')
+            
+            # move the three main output under main output folder from temp files
+            # if there is more than one contig and the prodigal outputs ahvn't been produces (first sample run)
+            
+            shutil.move(no_hits_genes_faa, config.output_folder+'/no_hits/' + 'no_hits.faa')
+            shutil.move(hits_genes_faa, config.output_folder+'/hits/' + 'hits.faa')
+            shutil.move(no_hits, config.output_folder+'/no_hits/' + 'no_hits.txt')
+            shutil.move(hits, config.output_folder+'/hits/' + 'hits.txt')
+            shutil.move(uniref_gene_map, config.output_folder+'/hits/' + 'uniref_gene_map.txt')
+            if not os.path.isfile(config.output_folder+"/prodigal.gff"):
+                shutil.move(genes_file_gff, config.output_folder+'/prodigal.gff')
+            if not os.path.isfile(config.output_folder+"/prodigal.faa"):
+                shutil.move(genes_file_faa, config.output_folder+'/prodigal.faa')
+            print ("Main output files for ppanini_press are written in: \n%s\n%s\n%s\n%s\n%s")% (config.output_folder+'/'+os.path.basename(os.path.normpath(abundance_file)),
+                                                config.output_folder+'/no_hits/' + 'no_hits.faa',
+                                                config.output_folder+'/no_hits/' + 'no_hits.txt',
+                                                config.output_folder+'/hits/' + 'hits.faa',
+                                                config.output_folder+'/hits/' + 'hits.txt',
+                                                config.output_folder+'/hits/' + 'uniref_gene_map.txt')
+ 
+    if not args.one_contig:
+        # move the three main output under main output folder from temp files
+        
+        shutil.move(no_hits_genes_faa, config.output_folder+'/no_hits/'+ config.file_basename+ '_no_hits.faa')
+        shutil.move(hits_genes_faa, config.output_folder+'/hits/'+ config.file_basename+ '_hits.faa')
+        shutil.move(no_hits, config.output_folder+'/no_hits/'+ config.file_basename+ '_no_hits.txt')
+        shutil.move(hits, config.output_folder+'/hits/'+ config.file_basename+ '_hits.txt')
+        shutil.move(uniref_gene_map, config.output_folder+'/hits/'+ config.file_basename+ '_uniref_gene_map.txt')
+        print ("Main output files for ppanini_press are written in: \n%s\n%s\n%s\n%s\n%s")% (config.output_folder+'/'+os.path.basename(os.path.normpath(abundance_file)),
+                                            config.output_folder+'/no_hits/'+ config.file_basename+ '_no_hits.faa',
+                                            config.output_folder+'/no_hits/'+ config.file_basename+ '_no_hits.txt',
+                                            config.output_folder+'/hits/'+ config.file_basename+ '_hits.faa',
+                                            config.output_folder+'/hits/'+ config.file_basename+ '_hits.txt',
+                                            config.output_folder+'/hits/'+ config.file_basename+ '_uniref_gene_map.txt')
         
         
 if __name__ == '__main__':
